Combine data-driven license tests in one module
